Normalization coefficients

+++ mmpretrain/engine/runners/multi_domain_train_loop.py
@@ -91,7 +91,7 @@ class MultiDomainTrainLoop(EpochBasedTrainLoop):
 
     # Initialize variables
     total_sum = torch.zeros(3)
-    total_sum_squared = torch.zeros(3)
+    total_sum_squared = torch.zeros(3, dtype=torch.float64)
     num_samples = 0
     num_pixels = 0
